Write custom encoding for line information
This encoding is optimized for the most common use-case in everyday code.
It's essentially run-length encoding in that we specify a line number and an
amount that it is repeated by, but there are two important aspects of this:

1.  The line number we encode in each segment is relative to the previous one.
    This makes sense given that, most of the time, code is fairly dense.
    It's very rare to have dozens of lines without any instructions,
    so why store the full line number each time? Just store the amount of lines
    since the last segment.
2.  A segment is only one byte long. That byte looks like this: CLLLAAAA.
    C = Continued, L = Line Offset, A = Amount.
    We use three bits to encode the line offset, giving us up to 8 values,
    and four bits to encode the amount of instructions on that line, giving us
    16 values. Whenever those are not enough, we simply set the Continued bit,
    and add another segment to the end of the list, which then contains the next
    3 and 4 bits of the line offset and amount respectively.

This all means that, in the optimal case of 15 instructions per line,
we only use 1 byte of line information per line of code.
Less than 15 instructions still uses one byte of course,
and 16 to 255 instructions uses two bytes, 256 to 4096 uses three, and so on.
Even a huge minified script with 16 million instructions on a single line would
only use 8 bytes on line information.

This all assumes that any two instructions in the bytecode have the same order
as in the source code (line offset can only be positive).
This isn't always the case and I haven't found a way to work around this yet.

The amount of segments that can be chained together is currently limited by the
size of the type that stores the final result (usize). Theoretically, this could
be infinite, but usize makes sense. (Who needs a script with more lines than
your CPU has memory addresses?)